Move and rename GeneratorVersion to GeneratorVersionProvider
